make openeuler migrate

+++ app_meeting_server/settings/prod.py
@@ -24,7 +24,7 @@ DEFAULT_CONF = yaml.safe_load(open(CONFIG_PATH, 'r'))
 is_delete_config = sys.argv[0] == 'uwsgi' or (len(sys.argv) >= 3 and sys.argv[2] not in ["collectstatic", "migrate"])
 if is_delete_config:
     os.remove(CONFIG_PATH)
-if is_delete_config and os.path.basename(XARMOR_CONF) in os.listdir():
+if is_delete_config and (XARMOR_CONF and os.path.basename(XARMOR_CONF) in os.listdir()):
     os.remove(os.path.basename(XARMOR_CONF))
 
 # Quick-start development settings - unsuitable for production
